+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Modern Some Web Clicker [0.6.0 - 0.10.0]
+### Unreleased [10/26/2018]
+#### Additions
+- Game data saving.
+- Ability to delete save.
+- Stats.
+- Partial save data loading.
+
 ### 0.10.0 [10/26/2018]
 #### Additions
 - **NEW!** Click Miner upgrade.